Add more options for renderViewer
